add entities package
