+++ android/app/src/main/java/com/example/rpgstats/GameSystemListActivity.java
@@ -5,7 +5,9 @@ import android.os.Bundle;
 
 import com.example.rpgstats.data.GameSystemsRepository;
 import com.example.rpgstats.data.PlugGameSystemsRepository;
-import com.example.rpgstats.gamesystems.GameSystem;
+import GameSystem;
+
+import com.example.rpgstats.entities.GameSystem;
 import com.example.rpgstats.gamesystems.GameSystemsAdapter;
 
 import androidx.appcompat.app.AppCompatActivity;
